Rename TaskManager to TaskList
This commit refactors the task management system to align with
the A-MoreOOP requirements. Specifically:

- Renamed TaskManager class to TaskList
- Updated all references to TaskManager across the project
- Adjusted package structure from taskManager to taskList

This change is part of the broader refactoring effort to make
the code more object-oriented and modular. It's a step towards
implementing the required class structure for the A-MoreOOP task.

Future commits will address the creation of Ui and Storage classes,
as well as further refactoring of the main Orion (soon to be Duke) class.
